feat: Refactor authentication router and add OAuth2 support
